Fleshed out and prettified power plant details page
