+++ src/components/state-details.js
@@ -2,15 +2,18 @@ import React, { useState } from "react"
 import { graphql, Link } from "gatsby"
 import Scrollspy from "react-scrollspy"
 import "react-bootstrap-table-next/dist/react-bootstrap-table2.min.css"
+
+import { slugToTitle } from "../helper-functions"
 import SingleBarChart from "../components/singlebar"
 import Layout from "../components/layout"
 import SEO from "../components/seo"
 import SimpleAreaChart from "../components/simpleareachart"
+import { getShortCitation } from "../constants/source-citations"
+import { getTerminologyHover } from "../constants/terminology-list"
+
 import AlreadyElectrifiedChart from "./already-electrified-chart"
 import PowerSourcesChart from "./power-sources-chart"
 import DisplayPlants from "./display-plants"
-import { getShortCitation } from "../constants/source-citations"
-import { getTerminologyHover } from "../constants/terminology-list"
 
 /**
  * Images - suffix with Img for clarity from actual JS files/variables
@@ -26,21 +29,6 @@ import PowerPlantMap from "../images/dirty-power-plants.jpg"
 
 // New images
 import DirtyPowerPlantImg from "../images/dirty-power-plant.png"
-
-const slugToTitle = placeName => {
-  const words = placeName.split("_")
-
-  for (let i = 0; i < words.length; i++) {
-    const word = words[i]
-    if (word === "of") {
-      words[i] = word
-    } else {
-      words[i] = word.charAt(0).toUpperCase() + word.slice(1)
-    }
-  }
-
-  return words.join(" ")
-}
 
 /**
  * Converts a very large number to a more readable string.
